Moved tests to relevant folder
