+++ src/test/unit/app/case/case-api-client.test.ts
@@ -2,11 +2,11 @@ import axios from 'axios';
 import type { AxiosInstance } from 'axios';
 import { LoggerInstance } from 'winston';
 
-import { UserDetails } from '../controller/AppRequest';
+import { UserDetails } from '../../../../main/app/controller/AppRequest';
 
-import { CaseApiClient, getCaseApiClient } from './case-api-client';
-import { CaseAssignedUserRole } from './case-roles';
-import { CaseRole } from './definition';
+import { CaseApiClient, getCaseApiClient } from '../../../../main/app/case/case-api-client';
+import { CaseAssignedUserRole } from '../../../../main/app/case/case-roles';
+import { CaseRole } from '../../../../main/app/case/definition';
 
 jest.mock('axios');
