+++ src/test/unit/app/case/case-api.test.ts
@@ -1,8 +1,8 @@
-import { UserDetails } from '../controller/AppRequest';
+import { UserDetails } from '../../../../main/app/controller/AppRequest';
 
-import { CaseApi, getCaseApi } from './case-api';
-import * as caseApiClient from './case-api-client';
-import { CaseRole } from './definition';
+import { CaseApi, getCaseApi } from '../../../../main/app/case/case-api';
+import * as caseApiClient from '../../../../main/app/case/case-api-client';
+import { CaseRole } from '../../../../main/app/case/definition';
 
 jest.mock('axios');
